refactor into a class, remove visibility var update from unload
