+++ src/answers-umd.js
@@ -32,8 +32,7 @@ import QueryUpdateListener from './core/statelisteners/queryupdatelistener';
 import { COMPONENT_REGISTRY } from './ui/components/registry';
 import { localizedDistance, parseLocale } from './core/utils/i18nutils';
 import createImpressionEvent from './core/analytics/createimpressionevent';
-import Searcher from './core/models/searcher';
-import { isIE } from './core/utils/useragent';
+import VisibilityAnalyticsHandler from './core/analytics/visibilityanalyticshandler';
 
 /** @typedef {import('./core/services/errorreporterservice').default} ErrorReporterService */
 /** @typedef {import('./core/services/analyticsreporterservice').default} AnalyticsReporterService */
@@ -272,7 +271,11 @@ class Answers {
 
       this.components.setAnalyticsReporter(this._analyticsReporterService);
       initScrollListener(this._analyticsReporterService);
-      this._initVisibilityChangeListeners(parsedConfig.search?.verticalKey);
+      const visibilityAnalyticsHandler = new VisibilityAnalyticsHandler(
+        this._analyticsReporterService,
+        parsedConfig.search?.verticalKey
+      );
+      visibilityAnalyticsHandler.initVisibilityChangeListeners();
     }
 
     this.core = new Core({
@@ -726,54 +729,6 @@ class Answers {
     } else {
       console.error(`Invalid visitor. Visitor was not set because "${JSON.stringify(visitor)}" does not have an id.`);
     }
-  }
-
-  /**
-   * Initialize visibility change event listener(s) to send analytics
-   * events when user close or navigate away from a result page.
-   */
-  _initVisibilityChangeListeners (verticalKey) {
-    /**
-     * Safari desktop listener and IE11 listeners fire visibility change event twice when switch
-     * to new tab and then close browser. This variable is used to ensure RESULTS_HIDDEN analytics event
-     * does not get send again if the page is already hidden.
-     */
-    let documentVisibilityState;
-    document.addEventListener('visibilitychange', () => {
-      if (document.visibilityState === 'hidden' && documentVisibilityState !== 'hidden') {
-        documentVisibilityState = 'hidden';
-        this._reportVisibilityChangeEvent(verticalKey, 'RESULTS_HIDDEN');
-      } else if (document.visibilityState === 'visible') {
-        documentVisibilityState = 'visible';
-        this._reportVisibilityChangeEvent(verticalKey, 'RESULTS_UNHIDDEN');
-      }
-    });
-
-    /**
-     * VisibilityChange API does not register when page is terminated (close tab/browser) in IE11.
-     * So, unload event is used to capture those RESULTS_HIDDEN scenarios.
-     */
-    if (isIE()) {
-      window.addEventListener('unload', () => {
-        if (documentVisibilityState !== 'hidden') {
-          documentVisibilityState = 'hidden';
-          this._reportVisibilityChangeEvent(verticalKey, 'RESULTS_HIDDEN');
-        }
-      });
-    }
-  }
-
-  /**
-   * Send visibility change related analytics event.
-   */
-  _reportVisibilityChangeEvent (verticalKey, eventName) {
-    const queryId = this._analyticsReporterService.getQueryId();
-    if (!queryId) {
-      return;
-    }
-    const searcher = verticalKey ? Searcher.VERTICAL : Searcher.UNIVERSAL;
-    const event = new AnalyticsEvent(eventName).addOptions({ searcher });
-    this._analyticsReporterService.report(event);
   }
 }
